some fixes for FreeBSD compilation
git-svn-id: http://svn.uvw.ru/mhddfs/trunk@699 d7d42532-9e51-4b1d-b199-9e4c94b77a30

+++ src/debug.c
@@ -36,7 +36,7 @@ int mhdd_debug(int level, const char *fmt, ...)
 		default:         fprintf(mhdd.debug, ": ");  break;
 	}
 
-	fprintf(mhdd.debug, "[%ld] ", pthread_self());
+	fprintf(mhdd.debug, "[%ld] ", (long int)pthread_self());
 
 	va_list ap;
 	va_start(ap, fmt);
